Add checks for existing symbolic links in symlink.sh

+++ symlink.sh
@@ -1,12 +1,15 @@
 # FIX ROOT SHELL
-! rm /root/.cshrc 2>/dev/null
-ln -s /vince/root/.cshrc /root/
-chsh -s /bin/csh root
-
+if [ ! -L "/root/.cshrc" ]; then
+	! rm /root/.cshrc 2>/dev/null
+	ln -s /vince/root/.cshrc /root/
+	chsh -s /bin/csh root
+fi
 
 
 
 # RE-ENABLE / FIX SSH
-! rm /etc/ssh/sshd_config 2>/dev/null
-ln -s /vince/etc/ssh/sshd_config /etc/ssh/
-service sshd restart
+if [ ! -L "/etc/ssh/sshd_config" ]; then
+	! rm /etc/ssh/sshd_config 2>/dev/null
+	ln -s /vince/etc/ssh/sshd_config /etc/ssh/
+	service sshd restart
+fi
